+++ Verizon/VerizonAdapterConfiguration.m
@@ -2,8 +2,7 @@
 #import <VerizonAdsCore/VerizonAdsCore.h>
 #import "VerizonAdapterConfiguration.h"
 
-NSString * const kMoPubVASAdapterVersion = @"1.1.4.2";
-NSString * const kMoPubVASNetworkSdkVersion = @"1.1.4";
+NSString * const kMoPubVASAdapterVersion = @"1.2.0.0";
 
 NSString * const kMoPubVASAdapterErrorWho = @"MoPubVASAdapter";
 NSString * const kMoPubVASAdapterPlacementId = @"placementId";
@@ -26,13 +25,19 @@ NSTimeInterval kMoPubVASAdapterSATimeoutInterval = 600;
 - (void)initializeNetworkWithConfiguration:(NSDictionary<NSString *, id> * _Nullable)configuration complete:(void(^ _Nullable)(NSError * _Nullable))complete
 {
     NSString *siteId = configuration[kMoPubVASAdapterSiteId];
-    if (siteId.length > 0 && [VASStandardEdition initializeWithSiteId:siteId])
-    {
-        MPLogInfo(@"VAS adapter version: %@", kMoPubVASAdapterVersion);
-    }
-    if (complete)
-    {
-        complete(nil);
+    if (siteId.length > 0) {
+        dispatch_async(dispatch_get_main_queue(), ^{
+            if ([VASStandardEdition initializeWithSiteId:siteId]) {
+                MPLogInfo(@"VAS adapter version: %@", kMoPubVASAdapterVersion);
+            }
+            if (complete) {
+                complete(nil);
+            }
+        });
+    } else {
+        if (complete) {
+            complete(nil);
+        }
     }
     
     if (MPLogging.consoleLogLevel == MPBLogLevelDebug) {
@@ -59,7 +64,21 @@ NSTimeInterval kMoPubVASAdapterSATimeoutInterval = 600;
 
 - (NSString *)networkSdkVersion
 {
-    return kMoPubVASNetworkSdkVersion;
+    NSString *editionName = [[[VASAds sharedInstance] configuration] stringForDomain:@"com.verizon.ads"
+                                                                                 key:@"editionName"
+                                                                         withDefault:nil];
+
+    NSString *editionVersion = [[[VASAds sharedInstance] configuration] stringForDomain:@"com.verizon.ads"
+                                                                                    key:@"editionVersion"
+                                                                            withDefault:nil];
+    if (editionName.length > 0 && editionVersion.length > 0) {
+        return [NSString stringWithFormat:@"%@-%@", editionName, editionVersion];
+    }
+    
+    NSString *adapterVersion = [self adapterVersion];
+    NSRange range = [adapterVersion rangeOfString:@"." options:NSBackwardsSearch];
+    
+    return adapterVersion.length > range.location ? [adapterVersion substringToIndex:range.location] : @"";
 }
 
 @end